// src/types/review-type.ts
export type PostingReviewType = {
  userName: string;
  advantage: string;
  disadvantage: string;
  review: string;
  rating: number;
  cameraId: number;
}

export type ReviewType = {
  id: string;
  userName: string;
  advantage: string;
  disadvantage: string;
  review: string;
  rating: number;
  createAt: string;
  cameraId: number;
}